[refactor] 멀티 모듈 적용에 따른 테스트 코드 리팩터링
