+++ build.gradle
@@ -44,6 +44,8 @@ project(':kitchenpos-menu') {
     dependencies {
         implementation project(':kitchenpos-common')
         implementation project(':kitchenpos-product')
+
+        testCompile project(':kitchenpos-product').sourceSets.test.output
     }
 }
 
@@ -51,6 +53,10 @@ project(':kitchenpos-order') {
     dependencies {
         implementation project(':kitchenpos-common')
         implementation project(':kitchenpos-menu')
+        implementation project(':kitchenpos-table')
+
+        testCompile project(':kitchenpos-menu').sourceSets.test.output
+        testCompile project(':kitchenpos-table').sourceSets.test.output
     }
 }
 
@@ -64,6 +70,5 @@ project(':kitchenpos-table') {
     dependencies {
         implementation project(':kitchenpos-common')
         implementation project(':kitchenpos-menu')
-        implementation project(':kitchenpos-order')
     }
 }
