+++ kitchenpos-menu/src/test/java/kitchenpos/domain/MenuProductFixture.java
@@ -1,6 +1,5 @@
 package kitchenpos.domain;
 
-import kitchenpos.menu.domain.MenuProduct;
 import org.springframework.test.util.ReflectionTestUtils;
 
 public class MenuProductFixture {
