+++ kitchenpos-menu/src/test/java/kitchenpos/domain/MenuProductTest.java
@@ -1,6 +1,5 @@
 package kitchenpos.domain;
 
-import kitchenpos.menu.domain.MenuProduct;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
