+++ kitchenpos-product/src/main/java/kitchenpos/Application.java
@@ -1,3 +1,5 @@
+package kitchenpos;
+
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
