+++ kitchenpos-product/src/main/java/kitchenpos/domain/ProductRepository.java
@@ -1,4 +1,4 @@
-package domain;
+package kitchenpos.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
